Interactivity: button-agnostic DoubleClickResponse
