+++ src/ScottPlot5/ScottPlot5/Interactivity/IMouseButtonAction.cs
@@ -3,4 +3,5 @@
 public interface IMouseButtonAction : IMouseAction
 {
     MouseButton Button { get; }
+    public bool IsPressed { get; }
 }
