+++ src/ScottPlot5/ScottPlot5/Interactivity/UserActions/MouseWheelUp.cs
@@ -3,6 +3,7 @@
 public record struct MouseWheelUp(Pixel Pixel) : IMouseButtonAction
 {
     public readonly MouseButton Button => StandardMouseButtons.Wheel;
+    public readonly bool IsPressed => true;
     public readonly string Device => $"mouse wheel";
     public readonly string Description => $"mouse wheel up";
     public DateTime DateTime { get; set; } = DateTime.Now;
